pyproject.toml
[project]
name = "a2a-langgraph-langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk[http-server]>=0.3.6",
    "aiohttp>=3.12.15",
    "asyncclick>=8.2.2.2",
    "chromadb>=1.1.0",
    "click>=8.3.0",
    "google-genai>=1.38.0",
    "httpx>=0.28.1",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.12",
    "langgraph>=0.6.7",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "pydantic>=2.11.9",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "streamlit>=1.49.1",
    "uvicorn>=0.36.0",
]
